Restructured spec and validation packages, fixed validatedspec tests.
