+++ mlte/value/types/integer.py
@@ -11,7 +11,7 @@ import typing
 from mlte.artifact.model import ArtifactHeaderModel, ArtifactModel
 from mlte.artifact.type import ArtifactType
 from mlte.evidence.metadata import EvidenceMetadata
-from mlte.validation.condition import Condition
+from mlte.spec.condition import Condition
 from mlte.validation.result import Failure, Success
 from mlte.value.artifact import Value
 from mlte.value.model import IntegerValueModel, ValueModel, ValueType
