+++ test/measurement/memory/test_local_process_memory_consumption.py
@@ -16,7 +16,7 @@ from mlte.measurement.memory import (
     LocalProcessMemoryConsumption,
     MemoryStatistics,
 )
-from mlte.validation.condition import Condition
+from mlte.spec.condition import Condition
 from mlte.validation.result import Failure, Success
 
 from ...support.meta import path_to_support
