+++ test/schema/test_report_schema.py
@@ -12,8 +12,8 @@ from jsonschema import ValidationError
 from mlte._private.schema import validate_report_schema
 from mlte.report import Dataset, Limitation, Report, UseCase, User
 from mlte.spec.spec import Spec
-from mlte.spec.validated_spec import ValidatedSpec
 from mlte.validation.result import Ignore
+from mlte.validation.validated_spec import ValidatedSpec
 
 
 @pytest.mark.skip("Pending artifact protocol implementation.")
