+++ test/validation/test_specvalidator.py
@@ -10,7 +10,7 @@ import pytest
 from mlte.evidence.metadata import EvidenceMetadata, Identifier
 from mlte.property.costs import StorageCost
 from mlte.spec.spec import Spec
-from mlte.spec.spec_validator import SpecValidator
+from mlte.validation.spec_validator import SpecValidator
 from mlte.value.types.integer import Integer
 
 
@@ -26,7 +26,7 @@ def test_no_property():
 
 
 def test_no_requirement():
-    # Spec validator does not have value for requirement.
+    # Spec validator does not have value for condition.
     spec = Spec(
         "spec", {StorageCost("rationale"): {"test": Integer.less_than(3)}}
     )
